% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perminf_code.R
\name{adjust_outcomes}
\alias{adjust_outcomes}
\title{Subtract off constant treatment effect from data}
\usage{
adjust_outcomes(data, Y, Z, tau0)
}
\value{
adjusted dataframe with new "Yadj" column.
}
\description{
Subtract off constant treatment effect from data
}
