% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_Quantile_20220919.R
\name{assign_CRE}
\alias{assign_CRE}
\title{Generate matrix of complete randomization assignments}
\usage{
assign_CRE(n, m, nperm)
}
\arguments{
\item{n}{Sample size}

\item{m}{Number units treated}

\item{nperm}{number of permutations.  Inf will generate all possible
permutations.}
}
\description{
Generate matrix of complete randomization assignments
}
