% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{data}
\name{benin}
\alias{benin}
\title{Benin Data}
\format{
A data frame with 24 rows and 7 variables:
\describe{
\item{Village}{Village ID}
\item{District}{District of village}libra
\item{Candidate}{Candidate in village}
\item{reg_voters}{Number of registered voters}
\item{VoteShare}{}
\item{vote_male}{}
\item{vote_male96}{}
\item{vote_female}{}
\item{vote_female96}{}
\item{Treatment}{Is Control, Clientelist, or Policy}
\item{Note}{Notes on our entering data from original paper.}
}
}
\usage{
benin
}
\description{
These are the small Benin dataset, taken from Table III.H of Watchekon (2002).
}
\keyword{datasets}
