% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{calc_CI_precision_finite}
\alias{calc_CI_precision_finite}
\title{Calculate properties of all confidence intervals for a
finite-sample dataset.}
\usage{
calc_CI_precision_finite(
  Y0,
  tau,
  p_tx,
  R = 100,
  k.vec = NULL,
  alpha = 0.05,
  alternative = "greater",
  method.list = list(name = "Stephenson", s = 10),
  score = NULL,
  stat.null = NULL,
  nperm = 1000,
  c = 0,
  quantile_n_CI = 0.5,
  summarise = TRUE
)
}
\arguments{
\item{R}{Number of permutation replicates}

\item{k.vec}{The vector of quantiles to investigate.  NULL means
all.}

\item{alpha}{Alpha for testing (default of 0.05)}

\item{alternative}{A character takes value "greater", "less" and
"two.sided", indicating the alternative hypothesis.}

\item{method.list}{A list specifies the choice of the rank sum test
statistic. For example, list(name="Wilcoxon") means the Wilcoxon
rank sum statistic, and list(name = "Stephenson", s = 10) means
the Stephenson rank sum statistic with parameter s=10.}

\item{score}{An \eqn{n} dimensional transformed ranks, i.e.,
(phi(1), phi(2), ..., phi(n)), where phi() denotes the rank
transformation function.}

\item{stat.null}{An vector whose empirical distribution
approximates the randomization distribution of the rank sum
statistic.}

\item{nperm}{A positive integer representing the number of
permutations for approximating the randomization distribution of
the rank sum statistic.}

\item{c}{Threshold for testing, if summarizing results.}

\item{quantile_n_CI}{What quantile of the distribution of number of
significant units should be reported?}

\item{summarise}{TRUE means summarise simulation results, FALSE
means return the raw matrix of CIs.}
}
\value{
If summarise = FALSE, Matrix with each column being a
simulation run and each row being a quantile, with the listed
confidence interval limit for each entry.  Otherwise a tibble
with a row for each quantile specified in k.vec.  The "n" column
is a repeated value of the median number of quantiles found
significant across simulations.
}
\description{
Given a fixed schedule of potential outcomes and a proportion to
assign to treatment, calculate (via simulation) the distribution of
all the confidence intervals.
}
\details{
This method will estimate (via simulation) the distributions all
simultaneous confidence intervals that can then be post processed
to determine their variability, and also the CI for the number of
positive units, etc.
}
\seealso{
calc_power_finite
}
