% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciresult.R
\name{ciresult}
\alias{ciresult}
\alias{arguments}
\alias{is.ciresult}
\alias{head.ciresult}
\alias{print.ciresult}
\alias{summary.ciresult}
\title{Result object for results of ci_quantile call}
\usage{
arguments(x)

is.ciresult(x)

\method{head}{ciresult}(x, n = 6L, ...)

\method{print}{ciresult}(x, ...)

\method{summary}{ciresult}(object, c = 0, k = NULL, ...)
}
\arguments{
\item{x}{a ciresult object
(except for is.ciresult, where it is a generic object to check).}

\item{n}{Number of rows to take}

\item{...}{extra options passed to print.ciresult}

\item{object}{object to summarize.}

\item{c}{Threshold for testing.}

\item{k}{Quantiles to report on.}
}
\value{
The parameters of the call

is.ciresult: TRUE if object is a ciresult object.

Data.frame of top rows

print: No return value; prints results.

summary: No return value; prints results.
}
\description{
The ciresult object is an S3 class that holds the results from \code{ci()}.

It has several methods that pull different information from this object, and
some printing methods for getting nicely formatted results.

Calculate some summary statistics on the result object.
}
