% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{data}
\name{electric_teachers}
\alias{electric_teachers}
\title{Teachers of Electric Circuits Professional Development}
\format{
A data frame with 233 rows and 7 variables:
\describe{
\item{\code{Site}}{character - Site ID (city where teachers were
recruited).  Randomization occured within site.}
\item{\code{T.ID}}{double - Teacher ID}
\item{\code{Tx}}{character -
Treatment group.  "D" is control.}
\item{\code{TxAny}}{double - Flag
for any treatment (main analysis did not find major differences of
PD approach.)}
\item{\code{Per.1}}{double - Baseline knowledge
(percent correct)}
\item{\code{Per.2}}{double - Endline knowledge
(percent correct)}
\item{\code{gain}}{double - Total gain from pre
to post.} }
}
\usage{
electric_teachers
}
\description{
This dataset describes the baseline and final test
scores on a test measuring the knowledge of electric circuits
given to teachers who were part of a randomized controlled trial
of a professional development program.

The trial had three treatment arms (different approaches to the
professional development) and a control arm.

The gain (post score - pre score) for all teachers is the change in the percentage point correct
on the test of electric circuit knowledge (we call this "content knowledge").

See Heller, J. I., Daehler, K. R., Wong, N., Shinohara, M., &
Miratrix, L. W. (2012). Differential effects of three professional
development models on teacher knowledge and student achievement in
elementary science. Journal of Research in Science Teaching,
49(3), 333–362. doi: 10.1002/tea.21004

See report on ERIC at https://files.eric.ed.gov/fulltext/ED514193.pdf
}
\details{
Not all treatments were given in a site; PD program was
delivered to classes of teachers.  Teachers individually
randomized to class, but there could be class (or PD teacher)
effects.
}
\keyword{datasets}
