% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{explore_stephenson_s}
\alias{explore_stephenson_s}
\title{Explore power across a range of s values}
\usage{
explore_stephenson_s(
  s = c(2, 5, 10, 30),
  n,
  Y0_distribution = rnorm,
  tx_function = "constant",
  p_tx = 0.5,
  R = 100,
  iter_per_set = 10,
  percentile = 1,
  alpha = 0.05,
  c = 0,
  k.vec = NULL,
  quantile_n_CI = 0.5,
  targeted_power = TRUE,
  alternative = "greater",
  nperm = 1000,
  parallel = FALSE,
  calc_ICC = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{s}{Numeric list of s values to test using the Stephenson rank.}

\item{k.vec}{Which quantiles to examine (if targeted_power =
FALSE).  See calc_CI_precision_finite.}

\item{targeted_power}{TRUE means focus on passed percentile and
calculate power for that specific quantile.  FALSE means
calculate simultenaous confidence intervals for all quantiles,
leaving power caculation to post-simulation exploration.}
}
\description{
Conduct power simulation against a specific data generating process to
see how power changes as a function of s in that scenario.
}
