% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{explore_stephenson_s_finite}
\alias{explore_stephenson_s_finite}
\title{Explore power and precision across a range of s values}
\usage{
explore_stephenson_s_finite(
  s = c(2, 5, 10, 30),
  Y0,
  tau,
  p_tx,
  R = 100,
  percentile = 1,
  k.vec = NULL,
  alpha = 0.05,
  c = 0,
  quantile_n_CI = 0.5,
  alternative = "greater",
  stat.nulls = NULL,
  nperm = 1000,
  targeted_power = TRUE
)
}
\arguments{
\item{s}{Numeric list of s values to test using the Stephenson
rank.}

\item{k.vec}{Which quantiles to examine (if targeted_power =
FALSE).  See calc_CI_precision_finite.}

\item{stat.nulls}{Optional list of the precomputed null
distributions of the statistics.}

\item{targeted_power}{TRUE means focus on passed percentile and
calculate power for that specific quantile.  FALSE means
calculate simultenaous confidence intervals for all quantiles,
leaving power caculation to post-simulation exploration.}
}
\description{
Conduct power simulation against specific finite-sample dataset to
see how power changes as a function of s.
}
