% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perminf_code.R
\name{find_ci_internal}
\alias{find_ci_internal}
\title{Find confidence interval threshold by looking for p-values crossing alpha.}
\usage{
find_ci_internal(
  Y,
  Z,
  S,
  coin_test,
  ci_level = 0.9,
  precision = 0.001,
  step = precision * 2^5,
  alternative,
  verbosity = 1,
  round_digits,
  distribution = "asymptotic",
  init_tau0 = NULL,
  tau_grid = NULL,
  ...
)
}
\arguments{
\item{step}{Initial step size for searching for threshold value.}
}
\description{
Utility function for actually doing the line search for the CI methods.
}
