% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{generate_finite_data}
\alias{generate_finite_data}
\title{Generate data for power simulation}
\usage{
generate_finite_data(
  n,
  Y0_distribution = rnorm,
  tx_function = tx_function_factory("constant")
)
}
\arguments{
\item{n}{Sample size to test}

\item{Y0_distribution}{Either a function to generate a set of Y0
values, or a list of discrete values to sample from (with
replacement).  Function takes single parameter of sample size.
Default is normal.}

\item{tx_function}{Function that takes Y0 and returns a set of
treatment effects.  Default is tx_constant.}
}
\value{
Tibble (dataframe) with two columns, one of Y0s and one
(tau) of treatment effects.
}
\description{
Given a desired sample size and a data generating mechanism,
generate a dataset of Y0s and individual treatment effects.
}
