% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perminf_code.R
\name{generate_four_stephenson_cis}
\alias{generate_four_stephenson_cis}
\title{Generate the four possible stephenson rank based confidence intervals.}
\usage{
generate_four_stephenson_cis(data, formula, tr_level = NULL, ...)
}
\arguments{
\item{data}{The data}

\item{formula}{The formula}

\item{...}{Parameters passed to \code{stephenson_test} and \code{find_ci_stephenson}}
}
\value{
Table of the four confidence intervals determined by which tail and
which alternative.
}
\description{
Generate the four possible stephenson rank based confidence intervals.
}
