% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fake_data.R
\name{make_fake_data}
\alias{make_fake_data}
\title{Make fake dataset for testing and illustration}
\usage{
make_fake_data(scenario = "D")
}
\arguments{
\item{scenario}{"A", "B", "C", or "D"; different kinds of data structure.}
}
\value{
Dataset with Y0, Y1, Yobs, and Z (treatment assignment).
}
\description{
These are all fixed set-size and -outcome datasets.  The pretend
randomization is perfectly balanced, so we have two units of each type with
one in treatment and one in control.
}
\details{
The scenarios are:

A: Long right tail, but also some very low treatment units. Clear evidence of
harm to some units and strong positive effects for others (the tx units are
in the tails, co units are interior)

B: Treatment has lowest values, control has highest.  It could be a simple
shift down (constant negative tx effect visually plausible).

C: Treatment helps in general, and helps some a lot.

D: Treatment helps those at the low end, but hurts those at the high end. So
testing for positive tx (alt = greater) on left tail should reject. Testing
for negative tx on right tail(alt = less) should also reject.
}
