% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_Quantile_20220919.R
\name{min_stat}
\alias{min_stat}
\title{minimum stat value in H_{k,c}}
\usage{
min_stat(Z, Y, k, c, method.list = NULL, score = NULL, ind.sort.treat = NULL)
}
\description{
minimum stat value in H_{k,c}
}
