% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simu_20220919.R
\name{normal_gen_PO}
\alias{normal_gen_PO}
\title{An example function for generating the potential outcomes}
\usage{
normal_gen_PO(n)
}
\arguments{
\item{n}{A numerical object specifying the sample size.}
}
\description{
A example function for generating the potential outcomes, which can be used for in the function simu_power for power comparison.
}
