% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_Quantile_20210801.R
\name{null_dist}
\alias{null_dist}
\title{rand dist of the rank score stat}
\usage{
null_dist(n, m, method.list = NULL, score = NULL, nperm = 10^5, Z.perm = NULL)
}
\description{
rand dist of the rank score stat
}
