% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile_effects.R
\name{plot_quantile_CIs}
\alias{plot_quantile_CIs}
\title{Plot quantile CIs}
\usage{
plot_quantile_CIs(result, k_start = NULL, main = NULL)
}
\arguments{
\item{result}{Result of a}

\item{k_start}{Quantile to start with (if NULL, plot all informative CIs).}

\item{main}{Title of plot}
}
\value{
The result object, again (invisibly).
}
\description{
Given the results of a quantile CI call, plot the results.
}
