% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perminf_code.R
\name{print.permconfint}
\alias{print.permconfint}
\title{Print method for results from confidence interval call.}
\usage{
\method{print}{permconfint}(x, ...)
}
\arguments{
\item{x}{Main argument.}

\item{...}{Additional arguments.}
}
\description{
This is a utility function.
}
