% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_bound_20220919.R
\name{pval_bound}
\alias{pval_bound}
\title{Randomization test for bounded null hypotheses}
\usage{
pval_bound(
  Z,
  Y,
  c = 0,
  method.list = list(name = "DIM"),
  alternative = "greater",
  Z.perm = NULL,
  nperm = 10^5,
  impute = "control"
)
}
\arguments{
\item{Z}{An \eqn{n} dimensional treatment assignment vector.}

\item{Y}{An \eqn{n} dimensional observed outcome vector.}

\item{c}{A scalar or vector specifying the bounded null hypothesis.}

\item{method.list}{A list specifies the choice of the rank sum test
statistic. For example, list(name = "DIM") means the
difference-in-means statistic, list(name="Wilcoxon") means the
Wilcoxon rank sum statistic, and list(name = "Stephenson", s =
10) means the Stephenson rank sum statistic with parameter s=10.}

\item{alternative}{A character takes value "greater" and "less"
indicating the direction of the alternative hypothesis.}

\item{Z.perm}{A \eqn{n \times nperm} matrix that specifies the
permutated assignments for approximating the null distribution of
the test statistic.}

\item{nperm}{A positive integer representing the number of
permutations for approximating the randomization distribution of
the test statistic}

\item{impute}{A character specifies the form of the test statistic.
"control" means we use the imputed control potential outcomes in
the test statistic, and "both" means we use the imputed observed
outcome in the test statistic.}
}
\value{
The p-value for testing the specified null hypothesis of
interest.
}
\description{
Obtain the p-value for testing the bounded null hypothesis H0: tau
<= c or H0: tau >= c
}
