% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_Quantile_20220919.R
\name{pval_quantile}
\alias{pval_quantile}
\title{Randomization test for quantiles of individual treatment effects}
\usage{
pval_quantile(
  Z,
  Y,
  k,
  c,
  alternative = "greater",
  method.list = list(name = "Stephenson", s = 10),
  score = NULL,
  stat.null = NULL,
  nperm = 10^6,
  Z.perm = NULL,
  switch = TRUE
)
}
\arguments{
\item{Z}{An \eqn{n} dimensional treatment assignment vector.}

\item{Y}{An \eqn{n} dimensional observed outcome vector.}

\item{k}{An integer between 1 and n specifying which quantile of
individual effect is of interest.}

\item{c}{A numerical object specifying the threshold for the null
hypothesis.}

\item{alternative}{A character takes value "greater", "less" and
"two.sided", indicating the alternative hypothesis.}

\item{method.list}{A list specifies the choice of the rank sum test
statistic. For example, list(name="Wilcoxon") means the Wilcoxon
rank sum statistic, and list(name = "Stephenson", s = 10) means
the Stephenson rank sum statistic with parameter s=10.}

\item{score}{An \eqn{n} dimensional transformed ranks, i.e.,
(phi(1), phi(2), ..., phi(n)), where phi() denotes the rank
transformation function.}

\item{stat.null}{An vector whose empirical distribution
approximates the randomization distribution of the rank sum
statistic.}

\item{nperm}{A positive integer representing the number of
permutations for approximating the randomization distribution of
the rank sum statistic.}

\item{Z.perm}{A \eqn{n \times nperm} matrix that specifies the
permutated assignments for approximating the null distribution of
the test statistic.}

\item{switch}{A logical object indicating whether performing
treatment label switching to make the treated group have larger
size.}
}
\value{
The p-value for testing the specified null hypothesis of
interest.
}
\description{
Obtain the p-value for testing the null hypothesis H0: tau_{(k)} <=
c, or H0: tau_{(k)} >= c, or H0: tau_{(k)} = c, where tau_{(k)}
denotes individual treatment effect at rank k.
}
