% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_Quantile_20210801.R
\name{rank_score}
\alias{rank_score}
\title{score function of the ranks}
\usage{
rank_score(n, method.list = list(name = "Wilcoxon"))
}
\description{
Calculate the scores of n units (ordered from low to high score) given a
specified test statistic, not allowing for ties.
}
