% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simu_20220919.R
\name{simu_power}
\alias{simu_power}
\title{Simulation for various Stephenson rank statistics}
\usage{
simu_power(
  gen = normal_gen_PO,
  n = 100,
  treat.prop = 0.5,
  iter.max = 100,
  Steph.s.vec = c(2, 5, 10),
  alternative = "greater",
  alpha = 0.05,
  nperm = 10^3,
  tol = 10^(-3),
  switch = TRUE,
  print.iter = 10
)
}
\arguments{
\item{gen}{A function that generates the potential outcomes.}

\item{n}{A numerical object specifying the sample size.}

\item{treat.prop}{A numerical object specifying the proportion of treated units.}

\item{iter.max}{A numerical object specifying the number of simulated data sets.}

\item{Steph.s.vec}{A vector specifying the parameters for a sequence of Stephenson rank statistics under investigation.}

\item{alternative}{A character takes value "greater" and "less", indicating the
direction of alternative hypotheses.}

\item{alpha}{A numerical object specifying the significance level.}

\item{nperm}{A positive integer representing the number of permutations for
approximating the randomization distribution of the rank sum statistic.}

\item{tol}{A numerical object specifying the precision of the obtained
confidence intervals. For example, if tol = 10^(-3), then the confidence
limits are precise up to 3 digits.}

\item{switch}{A logical object indicating whether performing treatment label
switching to make the treated group have larger size.}

\item{print.iter}{An integer determining how frequently we report the progress of the simulation.}
}
\value{
A list with the elements
\item{Steph.s.vec}{An vector specifying the parameters for a sequence of Stephenson rank
statistics under investigation.}
\item{dat.all}{A list containing all simulated data sets.}
\item{Z.CRE}{A \eqn{n} by iter.max matrix containing all simulated assignments.}
\item{method.list.all}{A list describing the Stephenson rank statistics under investigation.}
\item{ci.upper.all}{A list containing the inferred upper confidence limits for all quantiles of
individual effects for all simulated data sets using the specified Stephenson rank statistics.}
\item{ci.lower.all}{A list containing the inferred lower confidence limits for all quantiles of
individual effects for all simulated data sets using the specified Stephenson rank statistics.}
}
\description{
Conduct simulation for inferring quantiles of individual treatment effects using various
Stephenson rank statistics.
}
