% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perminf_code.R
\name{stephenson_rank}
\alias{stephenson_rank}
\title{Calculate stephenson ranks}
\usage{
stephenson_rank(y, subset_size = 6)
}
\arguments{
\item{y}{Vector of numerical values.}

\item{subset_size}{Look at all subsets of this size}
}
\value{
Vector of stephenson ranks for the given input vector of values.
}
\description{
For each observation, calculate number of subsets of size \code{subset_size} where
this observation is the largest in the set.
}
