% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perminf_code.R
\name{stephenson_test}
\alias{stephenson_test}
\title{Run the stephenson test.}
\usage{
stephenson_test(
  data,
  formula,
  distribution = approximate(1000),
  alternative = "less",
  tail = c("right", "left"),
  subset_size = 5,
  ...
)
}
\arguments{
\item{data}{The dataframe with the variables defined in 'formula'}

\item{formula}{Defines the testing relationship.  See coin package's
independence_test}

\item{distribution}{Passed to independence_test.  approximate(K) will do K
permutations.  "asymptotic" will calculate an asymptotic approximation.
"exact" will do exact.}

\item{alternative}{Direction of test.  E.g., "greater" will test whether the
first factor in the list of factors of treatment is higher than the other.
So the first factor level is the assumed treatment.}

\item{tail}{Do we focus on the right tail or left tail of the outcomes (i.e.,
do we flip the sign of the outcomes).}

\item{subset_size}{Look at all subsets of this size and score 1 if treatment
group has largest value.}

\item{...}{Extra arguments to pass to the coin's \code{independence_test()} method.}
}
\value{
The result of coin's \code{independence_test()}
}
\description{
This method runs coin's independence test with the stephenson rank.  It
basically is a shorthand call to coin.
}
