% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perminf_code.R
\name{stephenson_trafo}
\alias{stephenson_trafo}
\title{Bundle the stephenson rank function into a \code{transformation} that allows coin package to use it in its independence tests.}
\usage{
stephenson_trafo(data, subset_size = 6, sign = 1)
}
\arguments{
\item{data}{Dataframe}

\item{subset_size}{Size of subsets for seeing which observation is largest}

\item{sign}{Do we flip sign of all y to look for extreme low values rather than extreme high?}
}
\description{
Bundle the stephenson rank function into a \code{transformation} that allows coin package to use it in its independence tests.
}
