% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simu_20220919.R
\name{summary_power}
\alias{summary_power}
\title{Comparing the performance of various Stephenson rank statistics}
\usage{
summary_power(
  result,
  measure = c("CI_n(c)", "test_tau(k)", "CI_tau(k)"),
  alternative = "greater",
  cutoff = 0,
  k = NULL
)
}
\arguments{
\item{result}{A list containing the result from the simu_power function.}

\item{measure}{A character specifying how we compare different Stephenson rank statistics.
If it equals "CI_n(c)", then we compare the average confidence limits for the number of units with effects passing the threshold cutoff.
If it equals "CI_tau(k)", then we compare the average confidence limits for the individual treatment effect at rank k.
If it equals "test_tau(k)", then we compare the power for testing whether the individual treatment effect at rank k passes the threshold cutoff.}

\item{alternative}{A character takes value "greater" and "less", indicating the
direction of alternative hypotheses.}

\item{cutoff}{A numerical object specifying the cutoff.}

\item{k}{An integer that specifies the quantile of individual effects under investigation.}
}
\description{
Summarize the simulation to compare the performance of various Stephenson rank statistics.
}
