% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{summary_power_results}
\alias{summary_power_results}
\title{Summarize set of power simulation results}
\usage{
summary_power_results(
  result,
  alternative = "greater",
  c = 0,
  quantile_n_CI = 0.5
)
}
\arguments{
\item{result}{A tibble with one row for each quantile in the result object passed.}

\item{alternative}{A character takes value "greater" and "less",
indicating the direction of alternative hypotheses.}

\item{c}{A numerical object specifying the cutoff.}

\item{quantile_n_CI}{Quantile of the distribution of lower
confidence limits for the number of significant units.  Default
is 0.5 (median) reporting the median size of the confidence
interval for number of significant units.}
}
\description{
Summarize the simulation to compare the performance of various
Stephenson rank statistics.
}
\details{
Calculates three types of precision.  First, we compare the average
confidence limits for the number of units with effects passing the
threshold cutoff 'c' (default of 0). We also calculate the median
confidence limits for the individual treatment effect for all
ranks. Finally, we compare the power for testing whether the
individual treatment effect passes the threshold cutoff for each
rank.
}
