% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_Quantile_20210801.R
\name{conf_quant_larger}
\alias{conf_quant_larger}
\title{one-sided confidence interval (c, Inf) for all quantiles}
\usage{
conf_quant_larger(
  Z,
  Y,
  method.list = NULL,
  score = NULL,
  stat.null = NULL,
  nperm = 10^5,
  alpha = 0.05,
  tol = 10^(-3),
  ind.sort.treat = NULL
)
}
\arguments{
\item{Z}{treatment assignment (vector)}

\item{Y}{outcome (vector)}

\item{method.list}{choice of rank score stat, Default: NULL}

\item{score}{PARAM_DESCRIPTION, Default: NULL}

\item{stat.null}{PARAM_DESCRIPTION, Default: NULL}

\item{nperm}{PARAM_DESCRIPTION, Default: 10^5}

\item{alpha}{significance level, Default: 0.05}

\item{tol}{PARAM_DESCRIPTION, Default: 10^(-3)}

\item{ind.sort.treat}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
generate one-sided confidence intervals
}
\details{
DETAILS
}
