% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_Quantile_20220919.R
\name{conf_quant_larger}
\alias{conf_quant_larger}
\title{one-sided confidence interval (c, Inf) for all quantiles}
\usage{
conf_quant_larger(
  Z,
  Y,
  k.vec = NULL,
  method.list = NULL,
  score = NULL,
  stat.null = NULL,
  nperm = 10^5,
  Z.perm = NULL,
  alpha = 0.05,
  tol = 10^(-3),
  ind.sort.treat = NULL
)
}
\arguments{
\item{Z}{treatment assignment (vector)}

\item{Y}{outcome (vector)}
}
\description{
generate one-sided confidence intervals
}
