% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{explore_stephenson_s}
\alias{explore_stephenson_s}
\title{Explore power across a range of s values}
\usage{
explore_stephenson_s(
  s = c(2, 5, 10, 30),
  n,
  Y0_distribution = rnorm,
  tx_function = tx_function_factory("constant"),
  p_tx = 0.5,
  R = 100,
  iter_per_set = 10,
  percentile = 1,
  alpha = 0.05,
  c = 0,
  alternative = "greater",
  nperm = 1000,
  parallel = FALSE,
  calc_ICC = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{s}{Numeric list of s values to test using the Stephenson rank.}
}
\description{
Conduct power simulation against a specific data generating process to
see how power changes as a function of s in that scenario.
}
