% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{explore_stephenson_s_finite}
\alias{explore_stephenson_s_finite}
\title{Explore power and precision across a range of s values}
\usage{
explore_stephenson_s_finite(
  s = c(2, 5, 10, 30),
  Y0,
  tau,
  p_tx,
  R = 100,
  percentile = 1,
  k.vec = NULL,
  alpha = 0.05,
  c = 0,
  quantile_n_CI = NA,
  alternative = "greater",
  stat.nulls = NULL,
  nperm = 1000,
  targeted_power = TRUE
)
}
\arguments{
\item{s}{Numeric list of s values to test using the Stephenson
rank.}

\item{Y0}{Either a name of function to generate Y0s, or an
empirical distribution represented as a list of values.}

\item{p_tx}{Proportion of units to be treated.}

\item{R}{Number of simulation iterations to do.}

\item{percentile}{For targeted power, what percentile to target.}

\item{k.vec}{Which quantiles to examine (if targeted_power =
FALSE).}

\item{alpha}{Significance level of the test.}

\item{c}{Threshold bound for treatment impacts for hypothesis
testing.}

\item{quantile_n_CI}{What quantile of the distribution of lower
confidence limits for the number of significant units to
calculate.  NA (the default) means calculate mean instead of
median. 0.5 would report the median size of the confidence
interval for number of significant units.}

\item{alternative}{A character takes value "greater" and "less",
indicating the direction of alternative hypotheses.}

\item{stat.nulls}{Optional list of the precomputed null
distributions of the statistics.}

\item{targeted_power}{TRUE means focus on passed percentile and
calculate power for that specific quantile.  FALSE means
calculate simultenaous confidence intervals for all quantiles,
leaving power caculation to post-simulation exploration.}
}
\description{
Conduct power simulation against specific finite-sample dataset to
see how power changes as a function of s.
}
