% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{explore_stephenson_s_finite}
\alias{explore_stephenson_s_finite}
\title{Explore power across a range of s values}
\usage{
explore_stephenson_s_finite(
  s = c(2, 5, 10, 30),
  Y0,
  tau,
  p_tx,
  R = 100,
  percentile = 1,
  alpha = 0.05,
  c = 0,
  alternative = "greater",
  score = NULL,
  stat.null = NULL,
  nperm = 1000
)
}
\arguments{
\item{s}{Numeric list of s values to test using the Stephenson rank.}
}
\description{
Conduct power simulation against specific finite-sample dataset to
see how power changes as a function of s.
}
