% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{tx_function_factory}
\alias{tx_function_factory}
\title{Generate a function that makes treatment effects.}
\usage{
tx_function_factory(distribution, ATE = 0, tx_scale = 1, ...)
}
\arguments{
\item{distribution}{A string name of a distribution.  "constant" is
a constant treatment effect. "scale" is a scaling of Y0.  "rexp"
and "rnorm" are those random distributions.}

\item{ATE}{the average treatment effect.  Ignored by "scale"
option.}

\item{tx_scale}{Scaling factor to achieve desired standard
deviation of impacts generated by tx_function by (default is 1).
Ignored by "constant" and "scale" options.}

\item{...}{Extra parameters to pass to distribution function (e.g.,
sd for rnorm).}
}
\value{
A function of the form tx_function( Y0 ).
}
\description{
The function retuned is a function that returns a vector of
treatment impacts to go with passed vector of Y0 values.  I.e.,
final function is of format tx_function( Y0 ).
}
